+++ src/LPUARTSendHandler.cpp
@@ -95,8 +95,8 @@ void LPUARTSendHandler::irqHandler() {
   if ((control & LPUART_CTRL_TIE) != 0 && (status & LPUART_STAT_TDRE) != 0) {
     switch (sender_->state_) {
       case Sender::XmitStates::kBreak:
-        if (sender_->breakUseTimer_ &&
-            sender_->intervalTimer_.begin(
+        if (!sender_->breakUseTimer_ ||
+            !sender_->intervalTimer_.begin(
                 [&]() {
                   if (sender_->state_ == Sender::XmitStates::kBreak) {
                     port_->CTRL &= ~LPUART_CTRL_TXINV;
@@ -117,13 +117,15 @@ void LPUARTSendHandler::irqHandler() {
                   sender_->state_ = Sender::XmitStates::kData;
                   setActive();
                 },
-                sender_->adjustedBreakTime_)) {
-          // Invert the line as close as possible to the timer start
-          port_->CTRL |= LPUART_CTRL_TXINV;
-          setInactive();
-          sender_->breakStartTime_ = micros();
-        } else {
-          // Starting the timer failed, revert to the original way
+                sender_->breakTime_,
+                [&]() {
+                  // Invert the line as close as possible to the timer start
+                  port_->CTRL |= LPUART_CTRL_TXINV;
+                  setInactive();
+                  sender_->breakStartTime_ = micros();
+                })) {
+          // Not using a timer or starting it failed;
+          // revert to the original way
           breakSerialParams_.apply(port_);
           port_->DATA = 0;
           setCompleting();