+++ src/UARTSendHandler.cpp
@@ -86,8 +86,8 @@ void UARTSendHandler::irqHandler() {
   if ((control & UART_C2_TIE) != 0 && (status & UART_S1_TDRE) != 0) {
     switch (sender_->state_) {
       case Sender::XmitStates::kBreak:
-        if (sender_->breakUseTimer_ &&
-            sender_->intervalTimer_.begin(
+        if (!sender_->breakUseTimer_ ||
+            !sender_->intervalTimer_.begin(
                 [&]() {
                   if (sender_->state_ == Sender::XmitStates::kBreak) {
                     port_->C3 &= ~UART_C3_TXINV;
@@ -108,13 +108,15 @@ void UARTSendHandler::irqHandler() {
                   sender_->state_ = Sender::XmitStates::kData;
                   port_->C2 = UART_C2_TX_ACTIVE;
                 },
-                sender_->adjustedBreakTime_)) {
-          // Invert the line as close as possible to the timer start
-          port_->C3 |= UART_C3_TXINV;
-          port_->C2 = UART_C2_TX_INACTIVE;
-          sender_->breakStartTime_ = micros();
-        } else {
-          // Starting the timer failed, revert to the original way
+                sender_->breakTime_,
+                [&]() {
+                  // Invert the line as close as possible to the timer start
+                  port_->C3 |= UART_C3_TXINV;
+                  port_->C2 = UART_C2_TX_INACTIVE;
+                  sender_->breakStartTime_ = micros();
+                })) {
+          // Not using a timer or starting it failed;
+          // revert to the original way
           breakSerialParams_.apply(serialIndex_, port_);
           port_->D = 0;
           port_->C2 = UART_C2_TX_COMPLETING;