+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to
 ### Changed
 * No more volatile objects with weird copy constructors and
   assignment operators. Using atomic fences instead.
+* Replaced use of the Teensy library's `IntervalTimer` with a custom
+  `PeriodicTimer`. This improves BREAK time accuracy (when using a timer and not
+  serial parameters).
+  * This reverts the change from v4.1.0-beta.2.
 
 ## [4.1.1]